Algorithmic_Strategies/Perp_Futures_Funding_Arbitrage/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pyyaml>=6.0
python-binance>=1.0.17
ccxt>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0
